+++ src/main/java/com/epam/reportportal/cucumber/AbstractReporter.java
@@ -27,8 +27,8 @@ import com.epam.reportportal.service.item.TestCaseIdEntry;
 import com.epam.reportportal.service.tree.TestItemTree;
 import com.epam.reportportal.utils.*;
 import com.epam.reportportal.utils.files.ByteSource;
+import com.epam.reportportal.utils.formatting.MarkdownUtils;
 import com.epam.reportportal.utils.http.ContentType;
-import com.epam.reportportal.utils.markdown.MarkdownUtils;
 import com.epam.reportportal.utils.properties.SystemAttributesExtractor;
 import com.epam.ta.reportportal.ws.model.FinishExecutionRQ;
 import com.epam.ta.reportportal.ws.model.FinishTestItemRQ;
@@ -67,11 +67,11 @@ import java.util.stream.IntStream;
 import static com.epam.reportportal.cucumber.Utils.*;
 import static com.epam.reportportal.cucumber.util.ItemTreeUtils.createKey;
 import static com.epam.reportportal.cucumber.util.ItemTreeUtils.retrieveLeaf;
+import static com.epam.reportportal.utils.formatting.ExceptionUtils.getStackTrace;
 import static java.lang.String.format;
 import static java.util.Optional.of;
 import static java.util.Optional.ofNullable;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
-import static org.apache.commons.lang3.exception.ExceptionUtils.getStackTrace;
 
 /**
  * Abstract Cucumber 4.x formatter for Report Portal
@@ -92,7 +92,6 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	private static final String HOOK_ = "Hook: ";
 	private static final String DOCSTRING_DECORATOR = "\n\"\"\"\n";
 	private static final String ERROR_FORMAT = "Error:\n%s";
-	private static final String DESCRIPTION_ERROR_FORMAT = "%s\n" + ERROR_FORMAT;
 
 	public static final TestItemTree ITEM_TREE = new TestItemTree();
 	private static volatile ReportPortal REPORT_PORTAL = ReportPortal.builder().build();
@@ -115,11 +114,11 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	/**
 	 * This map uses to record the description of the scenario and the step to append the error to the description.
 	 */
-	private final Map<String, String> descriptionsMap = new ConcurrentHashMap<>();
+	private final Map<Maybe<String>, String> descriptionsMap = new ConcurrentHashMap<>();
 	/**
 	 * This map uses to record errors to append to the description.
 	 */
-	private final Map<String, Throwable> errorMap = new ConcurrentHashMap<>();
+	private final Map<Maybe<String>, Throwable> errorMap = new ConcurrentHashMap<>();
 
 	public static ReportPortal getReportPortal() {
 		return REPORT_PORTAL;
@@ -175,7 +174,8 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		if (definitionMatch != null) {
 			try {
 				Method method = retrieveMethod(definitionMatch);
-				return TestCaseIdUtils.getTestCaseId(method.getAnnotation(TestCaseId.class),
+				return TestCaseIdUtils.getTestCaseId(
+						method.getAnnotation(TestCaseId.class),
 						method,
 						codeRef,
 						(List<Object>) ARGUMENTS_TRANSFORM.apply(arguments)
@@ -248,14 +248,20 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	 * @param scenarioContext current scenario context
 	 */
 	protected void beforeScenario(RunningContext.FeatureContext featureContext, RunningContext.ScenarioContext scenarioContext) {
-		String scenarioName = Utils.buildName(scenarioContext.getKeyword(),
+		String scenarioName = Utils.buildName(
+				scenarioContext.getKeyword(),
 				AbstractReporter.COLON_INFIX,
 				scenarioContext.getTestCase().getName()
 		);
-		StartTestItemRQ startTestItemRQ = buildStartScenarioRequest(scenarioContext.getTestCase(), scenarioName, featureContext.getUri(), scenarioContext.getLine());
+		StartTestItemRQ startTestItemRQ = buildStartScenarioRequest(
+				scenarioContext.getTestCase(),
+				scenarioName,
+				featureContext.getUri(),
+				scenarioContext.getLine()
+		);
 		Maybe<String> id = startScenario(featureContext.getFeatureId(), startTestItemRQ);
 		scenarioContext.setId(id);
-		id.subscribe(scenarioId -> descriptionsMap.put(scenarioId, ofNullable(startTestItemRQ.getDescription()).orElse(StringUtils.EMPTY)));
+		descriptionsMap.put(id, ofNullable(startTestItemRQ.getDescription()).orElse(StringUtils.EMPTY));
 		if (launch.get().getParameters().isCallbackReportingEnabled()) {
 			addToTree(featureContext, scenarioContext);
 		}
@@ -273,9 +279,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 			return null;
 		} else {
 			if (STATUS_MAPPING.get(status) == null) {
-				LOGGER.error(String.format("Unable to find direct mapping between Cucumber and ReportPortal for TestItem with status: '%s'.",
-						status
-				));
+				LOGGER.error("Unable to find direct mapping between Cucumber and ReportPortal for TestItem with status: '{}'.", status);
 				return ItemStatus.SKIPPED;
 			}
 			return STATUS_MAPPING.get(status);
@@ -292,34 +296,35 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	 */
 	@Nonnull
 	@SuppressWarnings("unused")
-	protected Maybe<FinishTestItemRQ> buildFinishTestItemRequest(@Nonnull Maybe<String> itemId, @Nullable Date finishTime,
-																 @Nullable ItemStatus status) {
-		return itemId.map(id -> {
-			FinishTestItemRQ rq = new FinishTestItemRQ();
-			if (status == ItemStatus.FAILED) {
-				Optional<String> currentDescription = Optional.ofNullable(descriptionsMap.get(itemId.blockingGet()));
-				Optional<Throwable> currentError = Optional.ofNullable(errorMap.get(itemId.blockingGet()));
-				currentDescription.flatMap(description -> currentError
-								.map(errorMessage -> resolveDescriptionErrorMessage(description, errorMessage)))
-						.ifPresent(rq::setDescription);
-			}
-			ofNullable(status).ifPresent(s -> rq.setStatus(s.name()));
-			rq.setEndTime(finishTime);
-			return rq;
-		});
+	protected FinishTestItemRQ buildFinishTestItemRequest(@Nonnull Maybe<String> itemId, @Nullable Date finishTime,
+			@Nullable ItemStatus status) {
+		FinishTestItemRQ rq = new FinishTestItemRQ();
+		if (status == ItemStatus.FAILED) {
+			Optional<String> currentDescription = Optional.ofNullable(descriptionsMap.remove(itemId));
+			Optional<Throwable> currentError = Optional.ofNullable(errorMap.remove(itemId));
+			currentDescription.flatMap(description -> currentError.map(errorMessage -> resolveDescriptionErrorMessage(
+					description,
+					errorMessage
+			))).ifPresent(rq::setDescription);
+		}
+		ofNullable(status).ifPresent(s -> rq.setStatus(s.name()));
+		rq.setEndTime(finishTime);
+		return rq;
 	}
 
 	/**
 	 * Resolve description
+	 *
 	 * @param currentDescription Current description
-	 * @param error Error message
+	 * @param error              Error message
 	 * @return Description with error
 	 */
 	private String resolveDescriptionErrorMessage(String currentDescription, Throwable error) {
+		String errorStr = format(ERROR_FORMAT, getStackTrace(error, new Throwable()));
 		return Optional.ofNullable(currentDescription)
 				.filter(StringUtils::isNotBlank)
-				.map(description -> format(DESCRIPTION_ERROR_FORMAT, currentDescription, error))
-				.orElse(format(ERROR_FORMAT, error));
+				.map(description -> MarkdownUtils.asTwoParts(currentDescription, errorStr))
+				.orElse(errorStr);
 	}
 
 	/**
@@ -336,9 +341,9 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		}
 
 		Date endTime = Calendar.getInstance().getTime();
-		Maybe<FinishTestItemRQ> rqMaybe = buildFinishTestItemRequest(itemId, endTime, mapItemStatus(status));
+		FinishTestItemRQ rq = buildFinishTestItemRequest(itemId, endTime, mapItemStatus(status));
 		//noinspection ReactiveStreamsUnusedPublisher
-		rqMaybe.subscribe(rq -> launch.get().finishTestItem(itemId, rq));
+		launch.get().finishTestItem(itemId, rq);
 		return endTime;
 	}
 
@@ -367,8 +372,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		String featureUri = context.getFeatureUri();
 		currentScenarioContextMap.remove(Pair.of(context.getLine(), featureUri));
 		if (mapItemStatus(event.result.getStatus()) == ItemStatus.FAILED) {
-			Optional.ofNullable(event.result.getError())
-					.ifPresent(error -> context.getId().subscribe(id -> errorMap.put(id, error)));
+			Optional.ofNullable(event.result.getError()).ifPresent(error -> errorMap.put(context.getId(), error));
 		}
 		Date endTime = finishTestItem(context.getId(), event.result.getStatus());
 		featureEndTime.put(featureUri, endTime);
@@ -477,7 +481,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		String stepText = step.getText();
 		context.setCurrentText(stepText);
 		if (rq.isHasStats()) {
-			stepId.subscribe(id -> descriptionsMap.put(id, ofNullable(rq.getDescription()).orElse(StringUtils.EMPTY)));
+			descriptionsMap.put(stepId, ofNullable(rq.getDescription()).orElse(StringUtils.EMPTY));
 		}
 		if (launch.get().getParameters().isCallbackReportingEnabled()) {
 			addToTree(context, stepText, stepId);
@@ -492,9 +496,8 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	protected void afterStep(Result result) {
 		reportResult(result, null);
 		RunningContext.ScenarioContext context = getCurrentScenarioContext();
-		if (mapItemStatus(result.getStatus()) == ItemStatus.FAILED){
-			Optional.ofNullable(result.getError())
-					.ifPresent(error -> context.getCurrentStepId().subscribe(id -> errorMap.put(id, error)));
+		if (mapItemStatus(result.getStatus()) == ItemStatus.FAILED) {
+			Optional.ofNullable(result.getError()).ifPresent(error -> errorMap.put(context.getCurrentStepId(), error));
 		}
 		finishTestItem(context.getCurrentStepId(), result.getStatus());
 		context.setCurrentStepId(null);
@@ -518,8 +521,8 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	/**
 	 * Start before/after-hook item on Report Portal
 	 *
-	 * @param parentId  parent item id
-	 * @param rq hook start request
+	 * @param parentId parent item id
+	 * @param rq       hook start request
 	 * @return hook item id
 	 */
 	@Nonnull
@@ -598,7 +601,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		if (errorMessage != null) {
 			sendLog(errorMessage, level);
 		} else if (result.getError() != null) {
-			sendLog(getStackTrace(result.getError()), level);
+			sendLog(getStackTrace(result.getError(), new Throwable()), level);
 		}
 	}
 
@@ -623,7 +626,8 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		String type = ofNullable(mimeType).filter(ContentType::isValidType).orElseGet(() -> getDataType(data, name));
 		String attachmentName = ofNullable(name).filter(m -> !m.isEmpty())
 				.orElseGet(() -> ofNullable(type).map(t -> t.substring(0, t.indexOf("/"))).orElse(""));
-		ReportPortal.emitLog(new ReportPortalMessage(ByteSource.wrap(data), type, attachmentName),
+		ReportPortal.emitLog(
+				new ReportPortalMessage(ByteSource.wrap(data), type, attachmentName),
 				"UNKNOWN",
 				Calendar.getInstance().getTime()
 		);
@@ -667,9 +671,9 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 			return;
 		}
 		Date endTime = ofNullable(dateTime).orElse(Calendar.getInstance().getTime());
-		Maybe<FinishTestItemRQ> rqMaybe = buildFinishTestItemRequest(itemId, endTime, null);
+		FinishTestItemRQ rq = buildFinishTestItemRequest(itemId, endTime, null);
 		//noinspection ReactiveStreamsUnusedPublisher
-		rqMaybe.subscribe(rq -> launch.get().finishTestItem(itemId, rq));
+		launch.get().finishTestItem(itemId, rq);
 	}
 
 	private void removeFromTree(RunningContext.FeatureContext featureContext) {
@@ -793,14 +797,16 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		TestCase testCase = event.testCase;
 		RunningContext.FeatureContext newFeatureContext = new RunningContext.FeatureContext(testCase);
 		String featureUri = newFeatureContext.getUri();
-		RunningContext.FeatureContext featureContext = currentFeatureContextMap.computeIfAbsent(featureUri, u -> {
-			getRootItemId(); // trigger root item creation
-			newFeatureContext.setFeatureId(startFeature(buildStartFeatureRequest(newFeatureContext.getFeature(), featureUri)));
-			if (launch.get().getParameters().isCallbackReportingEnabled()) {
-				addToTree(newFeatureContext);
-			}
-			return newFeatureContext;
-		});
+		RunningContext.FeatureContext featureContext = currentFeatureContextMap.computeIfAbsent(
+				featureUri, u -> {
+					getRootItemId(); // trigger root item creation
+					newFeatureContext.setFeatureId(startFeature(buildStartFeatureRequest(newFeatureContext.getFeature(), featureUri)));
+					if (launch.get().getParameters().isCallbackReportingEnabled()) {
+						addToTree(newFeatureContext);
+					}
+					return newFeatureContext;
+				}
+		);
 
 		if (!featureContext.getUri().equals(testCase.getUri())) {
 			throw new IllegalStateException("Scenario URI does not match Feature URI.");
@@ -809,10 +815,12 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		RunningContext.ScenarioContext newScenarioContext = featureContext.getScenarioContext(testCase);
 
 		Pair<Integer, String> scenarioLineFeatureURI = Pair.of(newScenarioContext.getLine(), featureContext.getUri());
-		RunningContext.ScenarioContext scenarioContext = currentScenarioContextMap.computeIfAbsent(scenarioLineFeatureURI, k -> {
-			currentScenarioContext.set(newScenarioContext);
-			return newScenarioContext;
-		});
+		RunningContext.ScenarioContext scenarioContext = currentScenarioContextMap.computeIfAbsent(
+				scenarioLineFeatureURI, k -> {
+					currentScenarioContext.set(newScenarioContext);
+					return newScenarioContext;
+				}
+		);
 
 		beforeScenario(featureContext, scenarioContext);
 	}
@@ -840,14 +848,16 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	}
 
 	protected void addToTree(RunningContext.ScenarioContext scenarioContext, String text, Maybe<String> stepId) {
-		retrieveLeaf(scenarioContext.getFeatureUri(),
+		retrieveLeaf(
+				scenarioContext.getFeatureUri(),
 				scenarioContext.getLine(),
 				ITEM_TREE
 		).ifPresent(scenarioLeaf -> scenarioLeaf.getChildItems().put(createKey(text), TestItemTree.createTestItemLeaf(stepId)));
 	}
 
 	protected void removeFromTree(RunningContext.ScenarioContext scenarioContext, String text) {
-		retrieveLeaf(scenarioContext.getFeatureUri(),
+		retrieveLeaf(
+				scenarioContext.getFeatureUri(),
 				scenarioContext.getLine(),
 				ITEM_TREE
 		).ifPresent(scenarioLeaf -> scenarioLeaf.getChildItems().remove(createKey(text)));